[SG-107] feat: 단일 게시글 조회 api 연결 및 postId type 변경

+++ core/domain/src/main/kotlin/com/captures2024/soongan/core/domain/usecase/weekly/contests/GetPostInfoUseCase.kt
@@ -11,7 +11,7 @@ constructor(
     private val weeklyContestRepository: WeeklyContestRepository
 ) {
 
-    suspend operator fun invoke(postId: Int): Result<PostInfoDto> = runSuspendCatching {
+    suspend operator fun invoke(postId: Long): Result<PostInfoDto> = runSuspendCatching {
         return@runSuspendCatching weeklyContestRepository.getPostInfo(postId)
     }
 }
