+++ core/navigator/src/main/kotlin/com/captures2024/soongan/core/navigator/screen/main/home/HomeGalleryNavigator.kt
@@ -14,14 +14,14 @@ fun NavController.navigateToHomeGallery(navOptions: NavOptions?) = navigate(
     navOptions = navOptions,
 )
 
-fun NavController.setReportedPostId(postId: Int) {
+fun NavController.setReportedPostId(postId: Long) {
     this.previousBackStackEntry?.savedStateHandle?.set(savedStateHandleKey, postId)
 
     popBackStack<HomeGalleryNavigator>(inclusive = false)
 }
 
-fun NavController.getReportedPostId(): Int =
-    this.currentBackStackEntry?.savedStateHandle?.get(savedStateHandleKey) ?: -1
+fun NavController.getReportedPostId(): Long =
+    this.currentBackStackEntry?.savedStateHandle?.get(savedStateHandleKey) ?: -1L
 
 
 private const val savedStateHandleKey = "reported_post_id"
