+++ feature/profile/src/main/kotlin/com/captures2024/soongan/feature/profile/navigation/ProfileNavigation.kt
@@ -15,7 +15,7 @@ fun NavGraphBuilder.profile(
     navigateToBack: () -> Unit,
     navigateToEditProfile: () -> Unit,
     navigateToNotification: () -> Unit,
-    navigateToHomePost: (Int) -> Unit,
+    navigateToHomePost: (Long) -> Unit,
     navigateToRegistrationPost: () -> Unit,
     navigateToFAQ: () -> Unit,
 ) {
